字段类型调整

+++ plm-bom-configuration-domain/src/main/java/com/nio/ngfs/plm/bom/configuration/domain/model/configurationrule/ConfigurationRuleAggr.java
@@ -41,7 +41,7 @@ public class ConfigurationRuleAggr extends AbstractDo implements AggrRoot<Long> 
     /**
      * 创建Rule的目的，1-（Sales —> Eng），2-（Sales —> Sales），3-（Sales <—> Sales），4-（Sales X Sales），5-（Base Vehicle —> Sales）
      */
-    private Byte purpose;
+    private Integer purpose;
 
     /**
      * Rule的类型，Inclusive、Exclusive、Default、Available
